requirements.txt
rich>=13.0.0
questionary>=2.0.0
paramiko>=3.0.0
requests>=2.28.0
pyserial>=3.5
